refactor(client): extract logic from export files component
